fix: 스케쥴러 오류 수정

+++ src/main/kotlin/com/seogineer/kotlinspringlottogenerator/dto/LatestNumberResponse.kt
@@ -16,5 +16,5 @@ data class LatestNumberResponse (
     val bnusNo: Int,
     val firstWinamnt: BigInteger,
     val firstPrzwnerCo: Int,
-    val firstAccumamnt: Int,
+    val firstAccumamnt: BigInteger,
 )
